refactor: `BlockStdProvider` -> `BlockStdScope`
