+++ packages/block-std/src/index.ts
@@ -1,6 +1,6 @@
 export * from './command/index.js';
 export * from './event/index.js';
-export * from './provider/index.js';
+export * from './scope/index.js';
 export * from './selection/index.js';
 export * from './service/index.js';
 export * from './spec/index.js';
