+++ packages/block-std/src/scope/index.ts
