+++ packages/blocks/src/bookmark-block/components/bookmark-toolbar.ts
@@ -1,4 +1,4 @@
-import type { BlockStdProvider } from '@blocksuite/block-std';
+import type { BlockStdScope } from '@blocksuite/block-std';
 import { WithDisposable } from '@blocksuite/lit';
 import { flip, offset } from '@floating-ui/dom';
 import { css, html, LitElement, nothing } from 'lit';
@@ -57,7 +57,7 @@ export class BookmarkToolbar extends WithDisposable(LitElement) {
   moreButton!: HTMLElement;
 
   @property({ attribute: false })
-  std!: BlockStdProvider;
+  std!: BlockStdScope;
 
   private _moreMenuAbortController: AbortController | null = null;
